Refactored App.js: Separated AppBarComponent from App component.
